Create Bessel function in Julia
This code uses Julia to evaluate a boundary value problem using Bessel functions. The problems considers a circular disc with the temperature at the circumference held at 0 deg C and the disc is an an inital user defined temperature. The output is a graphical plot showing the temperautre of the disc at a given time and a given radius.